# Conflicts:
#	src/main/java/com/feature/gcoin/model/User.java

+++ src/main/java/com/feature/gcoin/security/TokenHelper.java
@@ -28,10 +28,10 @@ public class TokenHelper {
     public String SECRET;
 
     @Value("${jwt.expires_in}")
-    private int EXPIRES_IN;
+    private Long EXPIRES_IN;
 
     @Value("${jwt.mobile_expires_in}")
-    private int MOBILE_EXPIRES_IN;
+    private Long MOBILE_EXPIRES_IN;
 
     @Value("${jwt.header}")
     private String AUTH_HEADER;
@@ -138,7 +138,7 @@ public class TokenHelper {
         return new Date(timeProvider.now().getTime() + expiresIn * 1000);
     }
 
-    public int getExpiredIn(Device device) {
+    public Long getExpiredIn(Device device) {
         return device.isMobile() || device.isTablet() ? MOBILE_EXPIRES_IN : EXPIRES_IN;
     }
